Changed umount method in disk.sh. You can run umount any time.

+++ disk.sh
@@ -7,8 +7,10 @@ sudo mkdir -p $TMPDIR
 umount() {
     sudo umount ${MOUNT_DIR}
     sudo kpartx -d $TMPDIR/$IMAGE
-    [ -f $IMAGE ] && sudo rm $IMAGE
-    sudo cp $TMPDIR/$IMAGE $IMAGE
+	if [ -f $TMPDIR/$IMAGE ]; then
+		[ -f $IMAGE ] && sudo rm $IMAGE
+		sudo mv $TMPDIR/$IMAGE $IMAGE
+	fi
     sudo losetup -d /dev/loop[0-9] > /dev/null 2>&1 || return 0
 }
 
@@ -35,6 +37,7 @@ if [ $1 = "umount" ]; then
 fi
 
 if [ $1 = "grub-install" ]; then
+	umount
     [ -f $TMPDIR/$IMAGE ] || sudo rm $TMPDIR/$IMAGE
     sudo cp $IMAGE $TMPDIR/$IMAGE
     loopsetup
